+++ components/collector/src/source_collectors/jira/issue_status.py
@@ -11,7 +11,7 @@ class JiraIssueStatus(SourceCollector):
     async def _api_url(self) -> URL:
         """Override to get the issue, including the status field, from Jira."""
         url = await super()._api_url()
-        return URL(f"{url}/rest/api/2/issue/{self._issue_id}?fields=created,status,updated")
+        return URL(f"{url}/rest/api/2/issue/{self._issue_id}?fields=created,status,summary,updated")
 
     async def _landing_url(self, responses: SourceResponses) -> URL:
         """Override to add the issue to the landing URL."""
@@ -24,4 +24,5 @@ class JiraIssueStatus(SourceCollector):
         name = json["fields"]["status"]["name"]
         created = json["fields"]["created"]
         updated = json["fields"].get("updated")
-        return IssueStatus(self._issue_id, name=name, created=created, updated=updated)
+        summary = json["fields"].get("summary")
+        return IssueStatus(self._issue_id, name=name, created=created, updated=updated, summary=summary)