+++ components/frontend/src/measurement/IssueStatus.js
@@ -18,20 +18,27 @@ function IssuesWithoutTracker({ issueIds }) {
     return <>{issueIds.map((issueId) => <IssueWithoutTracker key={issueId} issueId={issueId} />)}</>
 }
 
+function labelDetails(issueStatus, issueTracker) {
+    let details = [<Label.Detail key="name">{issueStatus.name || "?"}</Label.Detail>]
+    if (issueStatus.summary && issueTracker?.parameters?.show_issue_summary) {
+        details.push(<Label.Detail key="summary">{issueStatus.summary}</Label.Detail>)
+    }
+    if (issueStatus.created && issueTracker?.parameters?.show_issue_creation_date) {
+        details.push(<Label.Detail key="created">Created <TimeAgo date={issueStatus.created} /></Label.Detail>)
+    }
+    if (issueStatus.updated && issueTracker?.parameters?.show_issue_update_date) {
+        details.push(<Label.Detail key="updated">Updated <TimeAgo date={issueStatus.updated} /></Label.Detail>)
+    }
+    return details
+}
+
 function IssueWithTracker({ issueStatus, issueTracker }) {
     let popupContent = "";  // Will contain error if any, otherwise creation and update dates, if any, else be empty
     if (issueStatus.connection_error) { popupContent = "Connection error" }
     if (issueStatus.parse_error) { popupContent = "Parse error" }
     const color = popupContent ? "red" : "blue";
     const basic = popupContent ? false : true;
-    let labelDetails = [<Label.Detail key="name">{issueStatus.name || "?"}</Label.Detail>]
-    if (issueStatus.created && issueTracker?.parameters?.show_issue_creation_date) {
-        labelDetails.push(<Label.Detail key="created">Created <TimeAgo date={issueStatus.created} /></Label.Detail>)
-    }
-    if (issueStatus.updated && issueTracker?.parameters?.show_issue_update_date) {
-        labelDetails.push(<Label.Detail key="updated">Updated <TimeAgo date={issueStatus.updated} /></Label.Detail>)
-    }
-    let label = <Label basic={basic} color={color}>{issueStatus.issue_id}{labelDetails}</Label>
+    let label = <Label basic={basic} color={color}>{issueStatus.issue_id}{labelDetails(issueStatus, issueTracker)}</Label>
     if (issueStatus.landing_url) {
         // Without the span, the popup doesn't work
         label = <span><HyperLink url={issueStatus.landing_url}>{label}</HyperLink></span>
@@ -43,7 +50,7 @@ function IssueWithTracker({ issueStatus, issueTracker }) {
         }
     }
     if (popupContent) {
-        label = <Popup content={popupContent} flowing hoverable trigger={label} />
+        label = <Popup header={issueStatus.summary} content={popupContent} flowing hoverable trigger={label} />
     }
     return label
 }